separate serializer code
