+++ flowback/comment/views.py
@@ -6,22 +6,7 @@ from rest_framework.views import APIView
 from flowback.comment.selectors import comment_list
 from flowback.comment.services import comment_create, comment_delete, comment_update
 from flowback.common.pagination import LimitOffsetPagination, get_paginated_response
-from flowback.files.serializers import FileSerializer
-
-
-class CommentOutputSerializer(serializers.Serializer):
-    id = serializers.IntegerField()
-    author_id = serializers.IntegerField()
-    author_name = serializers.CharField(source='author.username')
-    author_profile_image = serializers.ImageField(source='author.profile_image')
-    parent_id = serializers.IntegerField(allow_null=True)
-    created_at = serializers.DateTimeField()
-    edited = serializers.BooleanField()
-    active = serializers.BooleanField()
-    message = serializers.CharField(allow_null=True)
-    attachments = FileSerializer(source="attachments.filesegment_set", many=True, allow_null=True)
-    score = serializers.IntegerField()
-    num_replies = serializers.IntegerField()
+from flowback.comment.serializers import CommentOutputSerializer
 
 
 class CommentListAPI(APIView):
@@ -48,7 +33,7 @@ class CommentListAPI(APIView):
         score__lt = serializers.IntegerField(required=False)
 
     class OutputSerializer(CommentOutputSerializer):
-        pass
+        pass  # no changes
 
     def get(self, request, *args, **kwargs):
         serializer = self.FilterSerializer(data=request.query_params)
